requirements.txt
Flask==2.3.3
mysql-connector-python==8.2.0
Werkzeug==2.3.7
Pillow==10.0.1
python-dotenv==1.0.0
